requirements.txt
aprs==7.0.0
python-telegram-bot==12.2.0